Epic refactoring :D
