+++ src/pages/PathSelection.js
@@ -1,19 +1,8 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState, useEffect, useContext } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { QRCode } from 'react-qrcode-logo';
-import {
-  getAllPaths,
-  addPath,
-  deletePath,
-  getPathByName,
-  getWordsForPath,
-} from '../db/db';
-import {
-  connectToPeerAndReceive,
-  initializePeer,
-  sendDataOnConnection,
-} from '../utils/ShareUtils';
-import { importPath, exportPath } from '../utils/PathUtils';
+import { getAllPaths, getPathByName, getWordsForPath } from '../db/db';
+
+import { exportPath } from '../utils/PathUtils';
 import '../styles/PathSelection.css';
 import BackButton from '../components/universal/BackButton';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
@@ -21,32 +10,31 @@ import { faPlus } from '@fortawesome/free-solid-svg-icons';
 import EditButton from '../components/universal/EditButton';
 import DeleteButton from '../components/universal/DeleteButton';
 import ShareButton from '../components/universal/ShareButton';
-import QrScannerComponent from '../components/QrScannerComponent';
+import { PathContext } from '../components/pathSelection/PathContext';
+import AddPathModal from '../components/pathSelection/AddPathModal';
+import DeletePathModal from '../components/pathSelection/DeletePathModal';
+import PathNoWordsModal from '../components/pathSelection/PathNoWordsModal';
+import SharePathModal from '../components/pathSelection/SharePathModal';
+import ReceivePathModal from '../components/pathSelection/ReceivePathModal';
+import SharePathErrorModal from '../components/pathSelection/SharePathErrorModal';
 
 const PathSelection = () => {
   const navigate = useNavigate();
-  const [paths, setPaths] = useState([]);
-  const [newPath, setNewPath] = useState('');
-  const [isScanning, setIsScanning] = useState(false);
-  const [isScanningStarted, setIsScanningStarted] = useState(false);
+
+  const {
+    paths,
+    setPaths,
+    setCurrentPath,
+    setSharingSucceeded,
+    isSharingFailedModalOpen,
+    setIsSharingFailedModalOpen,
+  } = useContext(PathContext);
 
   const [isNewPathModalOpen, setIsNewPathModalOpen] = useState(false);
   const [isDeleteModalOpen, setIsDeleteModalOpen] = useState(false);
   const [isNoWordsInPathOpen, setIsNoWordsInPathOpen] = useState(false);
   const [isShareModalOpen, setIsShareModalOpen] = useState(false);
   const [isReceivePathModalOpen, setIsReceivePathModalOpen] = useState(false);
-  const [isSharingFailedModalOpen, setIsSharingFailedModalOpen] =
-    useState(false);
-
-  const [currentPath, setCurrentPath] = useState(null);
-
-  const [peer, setPeer] = useState(null);
-  const [peerId, setPeerId] = useState(null);
-  const [targetPeerIDInput, setTargetPeerIDInput] = useState('');
-  const [targetPeerID, setTargetPeerID] = useState(null);
-  const [sharingStarted, setSharingStarted] = useState(false);
-  const [sharingSucceeded, setSharingSucceeded] = useState(false);
-  const QRCODE_PREFIX = 'sanapolku:';
 
   // Fetch all paths from the database when the component loads
   useEffect(() => {
@@ -56,86 +44,6 @@ const PathSelection = () => {
       )
       .catch(() => console.error('Error fetching paths'));
   }, []);
-
-  useEffect(() => {
-    // Initialize WebRTC
-    const initPeer = async () => {
-      const { id: newId, peer: newPeer } = await initializePeer();
-      setPeerId(newId);
-      setPeer(newPeer);
-    };
-    if (!peer) {
-      initPeer();
-    }
-
-    return () => {
-      if (peer) {
-        peer.destroy();
-      }
-    };
-  }, [peer]);
-
-  useEffect(() => {
-    // Set up path to share
-    const handleNewConnection = async () => {
-      console.log('Setting up onConnection with', peer, currentPath);
-      sendDataOnConnection(peer, currentPath)
-        .then(() => {
-          setSharingSucceeded(true);
-          console.log('Successfully sent path', currentPath);
-        })
-        .catch((e) => {
-          closeShareModal();
-          openSharingFailedModal();
-          console.error('Connection failed', e);
-        });
-    };
-
-    if (!(peer && currentPath)) {
-      return;
-    }
-
-    handleNewConnection();
-  }, [peer, currentPath]);
-
-  const receivePath = async (id) => {
-    // Receive path from target
-    if (!peer) return;
-    try {
-      const importedPath = await connectToPeerAndReceive(peer, id, importPath);
-      const pathName = importedPath.name;
-      setPaths((prevPaths) => [...prevPaths, pathName]);
-      setSharingSucceeded(true);
-    } catch (e) {
-      closeReceivePathModal();
-      openSharingFailedModal();
-      console.error('Connection failed:', e);
-    } finally {
-      setSharingStarted(false);
-      setIsScanningStarted(false);
-    }
-  };
-
-  // Function to add a new path to the database and navigate
-  // to path management page
-  const handleAddPath = () => {
-    if (newPath.trim()) {
-      addPath(newPath)
-        .then(() => {
-          setPaths([...paths, newPath]);
-          setNewPath('');
-          console.log('Path added:', newPath);
-          setIsNewPathModalOpen(false);
-          navigate(`/muokkaapolkua/${newPath}`);
-        })
-        .catch((error) => {
-          console.error(error.message);
-          alert(error.message);
-        });
-    } else {
-      alert('Anna polulle nimi');
-    }
-  };
 
   // Function to navigate to the game
   const handlePathClick = async (path) => {
@@ -150,68 +58,10 @@ const PathSelection = () => {
     }
   };
 
-  // Navigate to path management page
-  const handleEditPathClick = (path) => {
-    navigate(`/muokkaapolkua/${path}`);
-    setIsNoWordsInPathOpen(false);
-  };
-
-  // Handle path deletion
-  const handlePathDelete = async () => {
-    if (!currentPath) return;
-
-    try {
-      const pathData = await getPathByName(currentPath); // Wait for getPathByName to resolve
-      if (!pathData || !pathData.id) {
-        console.error('Path not found:', currentPath);
-        alert('Path not found.');
-        return;
-      }
-
-      await deletePath(pathData.id);
-      setPaths((prevPaths) => prevPaths.filter((p) => p !== currentPath));
-      console.log(`Deleted path with name: ${currentPath}`);
-    } catch (error) {
-      console.error('Error deleting path:', error);
-      alert('Error deleting the path.');
-    }
-    setIsDeleteModalOpen(false);
-  };
-
-  const handleShareClick = () => {
-    setTargetPeerID(targetPeerIDInput);
-    receivePath(targetPeerIDInput);
-  };
-
-  const handleQRScan = async (scanResult) => {
-    const result = scanResult.data;
-    console.log(isScanningStarted);
-    if (result.startsWith(QRCODE_PREFIX) && !isScanningStarted) {
-      setIsScanningStarted(true);
-      result.substring();
-      const id = result.slice(QRCODE_PREFIX.length);
-      setTargetPeerID(id);
-      setIsScanning(false);
-      console.log('QR scanning ok');
-      console.log(targetPeerID);
-      setSharingStarted(true);
-      receivePath(id);
-    } else {
-      console.warn('Unknown QR code');
-      setIsScanningStarted(false);
-    }
-  };
-
   // Function to open the modal for deleting a path
   const openDeleteModal = (path) => {
     setCurrentPath(path);
     setIsDeleteModalOpen(true);
-  };
-
-  // Function to close the modal for deleting a path
-  const closeDeleteModal = () => {
-    setIsDeleteModalOpen(false);
-    setCurrentPath(null);
   };
 
   // Function to open the modal for creating a new path
@@ -219,22 +69,10 @@ const PathSelection = () => {
     setIsNewPathModalOpen(true);
   };
 
-  // Function to close the modal for creating a new path
-  const closeNewPathModal = () => {
-    setIsNewPathModalOpen(false);
-    setNewPath('');
-  };
-
   // Function to open the modal for informing lack of words in path
   const openNoWordsInPathModal = (path) => {
     setCurrentPath(path);
     setIsNoWordsInPathOpen(true);
-  };
-
-  // Function to close the modal for informing lack of words in path
-  const closeNoWordsInPathModal = () => {
-    setIsNoWordsInPathOpen(false);
-    setCurrentPath(null);
   };
 
   // Function to open the modal for sharing a path
@@ -244,35 +82,6 @@ const PathSelection = () => {
     });
     setIsShareModalOpen(true);
     setSharingSucceeded(false);
-  };
-
-  // Function to close the modal for sharing a path
-  const closeShareModal = () => {
-    setIsShareModalOpen(false);
-    setCurrentPath(null);
-  };
-
-  // Function to open the modal for sharing a path
-  const openReceivePathModal = () => {
-    setIsScanning(true);
-    setIsNewPathModalOpen(false);
-    setIsReceivePathModalOpen(true);
-  };
-
-  // Function to close the modal for sharing a path
-  const closeReceivePathModal = () => {
-    setIsReceivePathModalOpen(false);
-    setSharingSucceeded(false);
-  };
-
-  // Function to open the modal for instructions why path sharing failed
-  const openSharingFailedModal = () => {
-    setIsSharingFailedModalOpen(true);
-  };
-
-  // Function to close the modal for instructions why path sharing failed
-  const closeSharingFailedModal = () => {
-    setIsSharingFailedModalOpen(false);
   };
 
   return (
@@ -288,7 +97,6 @@ const PathSelection = () => {
           aria-label="Lisää uusi polku"
         />
       </div>
-
       {/* List of paths */}
       <div className="path-list">
         {paths.length > 0 ? (
@@ -318,162 +126,34 @@ const PathSelection = () => {
           <p className="no-paths">Ei polkuja.</p>
         )}
       </div>
-
-      {/* Modal for adding a new path */}
+      {/* Modal for adding a new path */};
       {isNewPathModalOpen && (
-        <div className="modal-overlay">
-          <div className="modal-content">
-            <h2>Vastaanota polku</h2>
-            <button
-              className="receive-path-button"
-              onClick={openReceivePathModal}
-            >
-              Siirry vastaanottamaan polku
-            </button>
-            <h2>Lisää uusi polku</h2>
-            <input
-              type="text"
-              value={newPath}
-              onChange={(e) => setNewPath(e.target.value)}
-              placeholder="Anna polun nimi"
-              className="modal-input"
-            />
-            <div className="modal-buttons">
-              <button className="cancel-button" onClick={closeNewPathModal}>
-                Peruuta
-              </button>
-              <button className="save-button" onClick={handleAddPath}>
-                Tallenna
-              </button>
-            </div>
-          </div>
-        </div>
+        <AddPathModal
+          onClose={() => setIsNewPathModalOpen(false)}
+          onOpenReceive={() => setIsReceivePathModalOpen(true)}
+        />
       )}
-
       {/* Modal for confirming deletion */}
       {isDeleteModalOpen && (
-        <div className="modal-overlay">
-          <div className="modal-content">
-            <h2>Vahvista poisto</h2>
-            <p>
-              Haluatko varmasti poistaa polun <b>{currentPath}</b>?
-            </p>
-            <div className="modal-buttons">
-              <button className="cancel-button" onClick={closeDeleteModal}>
-                Peruuta
-              </button>
-              <button className="save-button" onClick={handlePathDelete}>
-                Poista
-              </button>
-            </div>
-          </div>
-        </div>
+        <DeletePathModal onClose={() => setIsDeleteModalOpen(false)} />
       )}
-
       {/* Modal for informing user of lack of words in path */}
       {isNoWordsInPathOpen && (
-        <div className="modal-overlay">
-          <div className="modal-content">
-            <h2>Polulla ei ole vielä sanoja</h2>
-            <div className="modal-buttons">
-              <button className="save-button" onClick={closeNoWordsInPathModal}>
-                Palaa takaisin
-              </button>
-              <button
-                className="save-button"
-                onClick={() => handleEditPathClick(currentPath)}
-              >
-                Muokkaa polkua
-              </button>
-            </div>
-          </div>
-        </div>
+        <PathNoWordsModal onClose={() => setIsNoWordsInPathOpen(false)} />
       )}
-
       {/* Modal for sharing a path */}
       {isShareModalOpen && (
-        <div className="modal-overlay">
-          <div className="modal-content">
-            <h2>Polun jakaminen</h2>
-            {sharingSucceeded ? (
-              <label>Polun jakaminen onnistui!</label>
-            ) : (
-              <div>
-                <p>
-                  Näytä alla oleva QR-koodi polun vastaanottajalle. Jos kamera
-                  ei ole käytettävissä, polun jakaminen onnistuu QR-koodin alta
-                  löytyvän tunnisteen avulla.
-                </p>
-                <QRCode value={QRCODE_PREFIX + peerId} />
-                <span>Lähettäjän tunniste:</span>
-                <p>{peerId}</p>
-              </div>
-            )}
-            <button className="save-button" onClick={closeShareModal}>
-              Palaa takaisin
-            </button>
-          </div>
-        </div>
+        <SharePathModal onClose={() => setIsShareModalOpen(false)} />
       )}
-
       {/* Modal for sharing a path */}
       {isReceivePathModalOpen && (
-        <div className="modal-overlay">
-          <div className="modal-content">
-            <h2>Polun vastaanottaminen</h2>
-            {!sharingStarted ? (
-              <div>
-                {sharingSucceeded ? (
-                  <label>Polun jakaminen onnistui!</label>
-                ) : (
-                  <div>
-                    <p>
-                      Lue lähettäjän QR-koodi. Jos kamera ei ole käytettävissä,
-                      polun jakaminen onnistuu lähettäjän tunnisteen avulla.
-                    </p>
-                    {isScanning && (
-                      <QrScannerComponent onSuccess={handleQRScan} />
-                    )}
-                    <input
-                      className="fetch-input"
-                      type="text"
-                      value={targetPeerIDInput}
-                      placeholder="Lähettäjän tunniste"
-                      onChange={(e) => setTargetPeerIDInput(e.target.value)}
-                    />
-                    <button className="fetch-button" onClick={handleShareClick}>
-                      Hae polku
-                    </button>
-                  </div>
-                )}
-              </div>
-            ) : (
-              <label>Yhdistetään...</label>
-            )}
-            <button className="save-button" onClick={closeReceivePathModal}>
-              Palaa takaisin
-            </button>
-          </div>
-        </div>
+        <ReceivePathModal onClose={() => setIsReceivePathModalOpen(false)} />
       )}
-
       {/* Modal for path sharing error and instructions */}
       {isSharingFailedModalOpen && (
-        <div className="modal-overlay">
-          <div className="modal-content">
-            <h2>Polun jakaminen epäonnistui!</h2>
-            <p>
-              Yritä jakamista uudestaan. Jos jakaminen ei vieläkään onnistu,
-              ongelma voi johtua laitteiden käyttämistä verkoista. Siirtykää
-              käyttämään molemmilla laitteilla esimerkiksi samaa WIFI tai
-              mobiilidata verkkoa. Kun molemmat laitteet ovat yhdistettynä
-              samaan verkkoon, pitäisi polun jakamisen onnistua varmasti.
-            </p>
-            <button className="save-button" onClick={closeSharingFailedModal}>
-              Sulje
-            </button>
-          </div>
-        </div>
+        <SharePathErrorModal
+          onClose={() => setIsSharingFailedModalOpen(false)}
+        />
       )}
     </div>
   );
